% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt.R
\name{AdaPTGMM}
\alias{AdaPTGMM}
\title{Adaptive P-Value Thresholding for Multiple Hypothesis Testing using Gaussian Mixture Models}
\usage{
AdaPTGMM(
  x = NULL,
  p_values = NULL,
  z = NULL,
  testing = "one_sided",
  ndf = c(1, 3, 4),
  nclasses = c(2, 3, 4),
  niter = 5,
  alpha_m = 0.05,
  zeta = 0.1,
  lambda = 0.4,
  masking_shape = "tent",
  alphas = seq(0.01, 1, 0.01),
  selection = "BIC"
)
}
\arguments{
\item{x}{Vector of covariates TODO: (specify type)}

\item{p_values}{Vector of p-values (supply either p_values or test statistics)}

\item{z}{Vector of test statistics, TODO: not sure about how to deal with this yet}

\item{testing}{What form of testing procedure, either "\code{one_sided}" or "\code{interval}". Default is "\code{one_sided}".}

\item{ndf}{Vector of degrees of freedom in the spline basis for estimating \eqn{\beta}. Minimum degrees of freedom is 1,
representing no relationship between covariates and hypotheses. Note: Recommended to use <7 degrees of freedom.
Default is c(1,3,4). TODO: cannot use ndf=2 for now.}

\item{nclasses}{Vector of number of classes in Gaussian Mixture model. Minimum number of classes is 2.
Note: recommended to use <5 classes. Default is c(2,3,4).}

\item{niter}{Number of iterations per fitting of the expectation maximization algorithm.}

\item{alpha_m}{The maximum possible rejected p-value. We recommend \eqn{0.01\le \alpha_m \le 0.1}, default is 0.1.}

\item{zeta}{Controls minimum possible number of rejections, we recommend small values of zeta with low total number of samples.
If \code{zeta}=\eqn{\alpha} the desired FDR level, any number of rejections is possible.}

\item{lambda}{Controls where p-values are mirrored, boundary of blue region. TODO: Fix wording. We recommend \eqn{0.3\le\lambda\le 0.5}, default is 0.4
This is the most expensive part of the procedure, we recommend smaller number (<5) of iterations for larger problems. Default is 10.}

\item{masking_shape}{Controls the shape of the masking function, either "\code{tent}" or "\code{comb}" masking functions. Default is "\code{tent}".}

\item{alphas}{Vector of FDR levels of interest. Default is [0.9,0.89,...,0.02,0.01].}

\item{selection}{Type of selection procedure in model_selection. Options include "\code{BIC}", "\code{AIC}", "\code{cross_validation}". Default is "\code{BIC}".}
}
\description{
Fits a Gaussian Mixture model to the distribution of test statistics and
returns rejections and fitted parameters.
}
\details{
The constraint on these masking function parameters is
\deqn{0< \alpha_m \le \lambda <\lambda+ \alpha_m/\zeta\le 1.}
Setting \code{alpha_m} to 0.5, \code{lambda} to 0.5, \code{zeta} to 1, and \code{masking_shape} to "\code{tent}" results in the AdaPT default masking function.
}
