% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilities.R
\name{class_prob}
\alias{class_prob}
\title{Compute Class Probabilities}
\usage{
class_prob(beta, nclasses)
}
\arguments{
\item{beta}{fitted multinom}

\item{nclasses}{number of Gaussian mixture classes}
}
\value{
Dataframe of class probabilities, columns represent classes and rows represent each hypothesis
}
\description{
Compute Class Probabilities
}
