% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt.R
\name{compute_fdphat}
\alias{compute_fdphat}
\title{Compute FDPHat}
\usage{
compute_fdphat(data, args)
}
\arguments{
\item{data}{data for model}

\item{args}{args for model}
}
\value{
List with A_t, R_t, and estimated FDPHat
}
\description{
Compute FDPHat
}
\details{
Computes FDPHat with (A_t+1)/(max(R_t,1))
}
