% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{EM}
\alias{EM}
\title{Perform Expectation Maximization Algorithm on Gaussian Mixture Model}
\usage{
EM(model, preset_iter = NULL)
}
\arguments{
\item{model}{Model class with data, args, and initialize parameters}
}
\value{
Model class with updated parameters.
}
\description{
Will perform \code{niter} iterations of the expectation maximization procedure.
This procedure estimates \code{gammas}, the probabilities for each class, then uses these to fit \eqn{\beta}.
With the updated beta, data, and parameters, we estimate the probability of a class and true p-value based on
the covariate and masked value.
}
