% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilities.R
\name{prob_conditioned_on_class}
\alias{prob_conditioned_on_class}
\title{Compute the probability of big/small and masked p-value conditioned on class}
\usage{
prob_conditioned_on_class(model)
}
\arguments{
\item{model}{}
}
\value{
Dataframe, columns correspond to P[a_i=a,\tilde p_i | \gamma=k]P[\gamma=k | x_i],
divided by \zeta if a corresponds to big p-values
}
\description{
Compute the probability of big/small and masked p-value conditioned on class
}
