% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilities.R
\name{prob_jacobian}
\alias{prob_jacobian}
\title{Helper function to compute probability of big/small and masked p-value conditioned on class
for specific hypothesis}
\usage{
prob_jacobian(z, mean, var)
}
\arguments{
\item{z}{Test statistic value}

\item{mean}{mean of Gaussian mixture model class k}

\item{var}{variance of Gaussian mixture model class k}
}
\value{
P[a_i=a,\tilde p_i | \gamma=k]
}
\description{
Helper function to compute probability of big/small and masked p-value conditioned on class
for specific hypothesis
}
\details{
Computation is equivalent to phi(z,mean,var)/phi(z,0,1) where phi(z,mu,tau) is the density of a Gaussian
random variable with mean mu and variance tau at z.
}
