% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masking.R
\name{select_masking_params}
\alias{select_masking_params}
\title{Select masking parameters}
\usage{
select_masking_params(n, alpha_m, zeta, lambda)
}
\description{
zeta influences the minimum possible number of rejections. The minimum number of possible rejections
at FDR level \code{alpha} is \code{1/(\zeta\alpha)}. Thus for \code{alpha}=0.05, this corresponds to
\code{20/\zeta}.
}
