% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt.R
\name{adapt_gmm}
\alias{adapt_gmm}
\title{Adaptive P-Value Thresholding for Multiple Hypothesis Testing using Gaussian Mixture Models}
\usage{
adapt_gmm(
  x = NULL,
  pvals = NULL,
  z = NULL,
  testing = "one_sided",
  rendpoint = NULL,
  lendpoint = NULL,
  beta_formulas = NULL,
  model_type = "glm",
  nclasses = c(2, 3, 4),
  niter_fit = 5,
  niter_ms = 10,
  nfits = 5,
  alpha_m = NULL,
  zeta = NULL,
  lambda = NULL,
  masking_shape = "tent",
  alphas = seq(0.01, 1, 0.01),
  cr = "AIC",
  tol = 1e-04,
  initialization = "kmeans",
  intercept_model = TRUE,
  return_all_models = FALSE
)
}
\arguments{
\item{x}{Dataframe of covariates TODO: (specify type)}

\item{pvals}{Vector of p-values (supply either pvals or test statistics)}

\item{z}{Vector of test statistics, required if \code{testing}='\code{interval}'.}

\item{testing}{The form of testing procedure, either "\code{one_sided}" or "\code{interval}". Default is "\code{one_sided}".}

\item{rendpoint}{Corresponds to right endpoint of null hypothesis interval. Required if \code{testing}=\code{'interval'}.}

\item{lendpoint}{Corresponds to left endpoint of null hypothesis interval. If interval testing and \code{lendpoint} is blank,
\code{lendpoint} will be assumed to be \code{-rendpoint}.}

\item{beta_formulas}{TODO: Fill this in}

\item{model_type}{Type of model used for modeling beta, options include \code{gam} and \code{glm}. Default is \code{glm}.}

\item{nclasses}{Vector of number of classes in Gaussian Mixture model. The vector corresponds to the possible
number of classes to select in the model selection procedure. Minimum number of classes is 2.
Note: recommended to use <5 classes. Default is c(2,3,4). The greater the number of degrees of freedom the longer it takes the EM procedure to fit, and the
longer the list of possible values, the longer the model selection procedure takes.}

\item{niter_fit}{Number of iterations of EM per model update.}

\item{niter_ms}{Number of iterations of EM in model selection.}

\item{nfits}{Number of model fitting steps.}

\item{alpha_m}{The maximum possible rejected p-value. We recommend \eqn{0.01\le \alpha_m \le 0.1}, default is 0.1.}

\item{lambda}{Controls where p-values are mirrored, boundary of blue region. TODO: Fix wording. We recommend \eqn{0.3\le\lambda\le 0.5}, default is 0.4
This is the most expensive part of the procedure, we recommend smaller number (<5) of iterations for larger problems. Default is 10.}

\item{masking_shape}{Controls the shape of the masking function, either "\code{tent}" or "\code{comb}" masking functions. Default is "\code{tent}".}

\item{alphas}{Vector of FDR levels of interest. Default is [0.01,0.02,...,0.89,0.9].}

\item{cr}{Type of selection criterion in model_selection. Options include "\code{BIC}", "\code{AIC}", "\code{AICC}", \code{HIC}. Default is "\code{aIC}".}

\item{tol}{Positive scalar for early stopping if mu and tau do not update by more than \code{tol}.}

\item{initialization}{Type of initialization procedure to use. Options include "\code{kmeans}" or "\code{random}", where
"\code{random}" corresponds to drawing uniformly from predefined intervals. Default is "\code{kmeans}".}

\item{intercept_model}{Boolean.Include intercept only model in the model selection, default is \code{TRUE}.}

\item{return_all_models}{Boolean, whether to return all models used at various alpha levels. Default \code{FALSE}.
Required \code{TRUE} for plot_nn_masking. Warning, can be expensive to store all models for large problems.}

\item{verbose}{Boolean. Include print statements at each stage of the procedure.}
}
\description{
Fits a Gaussian Mixture model to the distribution of test statistics and
returns rejections and fitted parameters.
}
\details{
The constraint on these masking function parameters is
\deqn{0< \alpha_m \le \lambda <\lambda+ \alpha_m\zeta\le 1.}
Setting \code{alpha_m} to 0.5, \code{lambda} to 0.5, \code{zeta} to 1, and \code{masking_shape} to "\code{tent}" results in the AdaPT masking function.
}
