% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{model_selection}
\alias{model_selection}
\title{Model selection for degrees of freedom and number of classes}
\usage{
model_selection(
  data,
  args,
  beta_formulas,
  nclasses_list,
  cr,
  initialization,
  param_grid = NULL
)
}
\arguments{
\item{data}{data of model}

\item{args}{args of model}

\item{beta_formulas}{List of formulas for beta expansion}

\item{nclasses_list}{Vector for possible number of classes}

\item{cr}{Criterion for model selection}

\item{initialization}{Initialization procedure, either kmeans or random}
}
\value{
Initialized and pretrained model with best performance based on selection criterion
}
\description{
Model selection for degrees of freedom and number of classes
}
\keyword{Internal}
