% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{model_selection}
\alias{model_selection}
\title{Model selection for degrees of freedom and number of classes}
\usage{
model_selection(
  data,
  args,
  ndf_list,
  nclasses_list,
  selection,
  training_proportion = 0.6
)
}
\arguments{
\item{data}{data of model}

\item{args}{args of model}

\item{ndf_list}{Vector for possible degrees of freedom}

\item{nclasses_list}{Vector for possible number of classes}

\item{selection}{Criteria for model selection}

\item{training_proportion}{Proportion of data used for training, in the \code{selection}='\code{cross_validation}' case.}
}
\description{
Model selection for degrees of freedom and number of classes
}
\keyword{Internal}
